move test ofg image pair likelihojod up one
